app/business-api/python/transaction/pyproject.toml
[project]
name = "banking-assistant-transaction-mcp"
version = "0.1.0"
description = "Python MCP server for transactions business-api using fastmcp"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0",
    "fastapi[all]==0.116.1",
    "uvicorn",
    "azure-monitor-opentelemetry>=1.6.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0"
]
